pyproject.toml
[project]
name = "imagefolderpicker"
version = "2.0.0"
description = "A ComfyUI node for browsing and selecting images from a folder with thumbnail preview"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0",
    "watchdog>=4.0.0",
]

[project.urls]
Repository = "https://github.com/yourusername/ComfyUI-ImageFolderPicker"

[tool.comfy]
PublisherId = "your-publisher-id"
DisplayName = "Image Folder Picker"
Icon = "📁"
